requirements.txt
streamlit==1.28.1
langchain==0.1.5
langchain-core==0.1.16
langchain-openai==0.1.0
langgraph==0.0.42
python-dotenv==1.0.0
yfinance==0.2.37
pandas==2.2.2
